+++ mod_pow.h
@@ -1,14 +1,15 @@
 #pragma once
 
 
+
 ///////////////////////////////
 //// MOD固定. 少し速い (E.g. 2000ms --> 1900ms)
 
 //O(logn)
 //return x**n % MOD
-ll mod_pow(ll x, ll n){
+long long mod_pow(long long x, long long n){
     x %= MOD;
-    ll res = 1;
+    long long res = 1;
     while(n > 0){
         if(n & 1) res = res * x % MOD;
         x = x * x % MOD;
@@ -18,8 +19,8 @@ ll mod_pow(ll x, ll n){
 }
 
 //return x / y % MOD
-ll mod_div(ll x, ll y){
-    ll ret = x * mod_pow(y, MOD-2);
+long long mod_div(long long x, long long y){
+    long long ret = x * mod_pow(y, MOD-2);
     return ret % MOD;
 }
 
@@ -28,9 +29,9 @@ ll mod_div(ll x, ll y){
 
 //O(logn)
 //return x**n % mod
-ll mod_pow(ll x, ll n, ll mod){
+long long mod_pow(long long x, long long n, long long mod){
     x %= mod;
-    ll res = 1;
+    long long res = 1;
     while(n > 0){
         if(n & 1) res = res * x % mod;
         x = x * x % mod;
@@ -40,7 +41,7 @@ ll mod_pow(ll x, ll n, ll mod){
 }
 
 //return x / y % MOD
-ll mod_div(ll x, ll y, ll mod){
-    ll ret = x * mod_pow(y, mod-2, mod);
+long long mod_div(long long x, long long y, long long mod){
+    long long ret = x * mod_pow(y, mod-2, mod);
     return ret % mod;
 }
