+++ salesman.h
@@ -1,17 +1,17 @@
 #pragma once
 
 
-const ll INF = 1e15;
+const long long INF = 1e15;
 #define MAX_N 18
 
-ll N;
-ll D[MAX_N][MAX_N];
+long long N;
+long long D[MAX_N][MAX_N];
 
-ll DP[1 << MAX_N][MAX_N];
+long long DP[1 << MAX_N][MAX_N];
 
 // すべての頂点を一度ずつめぐり帰ってくる経路のうち、重みの総和の最小値を求める。
 // O(2**n n**2)
-ll rec(ll S, ll v){
+long long rec(long long S, long long v){
     if(DP[S][v] >= 0){
         return DP[S][v];
     }
@@ -19,7 +19,7 @@ ll rec(ll S, ll v){
     if(S == (1 << N)-1 && v == 0)
         return DP[S][v] = 0;
 
-    ll res = INF;
+    long long res = INF;
     for(int u = 0; u < N; u++){
         if(!(S >> u & 1)){
             res = min(res, rec(S | 1 << u, u) + D[v][u]);
@@ -35,4 +35,4 @@ REP(i, 1<<MAX_N){
     }
 }
 
-ll ans = rec(0, 0);
+long long ans = rec(0, 0);
